// src/app/shared/interfaces/login.interface.ts
export interface Login {
  UserName: string;
  Password: string;
}


// Generated by https://quicktype.io

export interface LoginResponse {
  user: User;
  token: string;
}

export interface User {
  name: string;
  lastName: string;
  email: string;
  userName: string;
  rolId: number;
}


export interface RegisterResponse {
  name: string,
  lastName: string,
  email: string,
  password: string
}